Some minor stuff

+++ app/Model/Image.php
@@ -2,10 +2,16 @@
 class Image extends AppModel {
 	public $belongsTo = array('Event');
 	
-	public function remove($id = null) {
-		$this->delete($id);
-		$file = APP . 'webroot' . DS . 'img' . DS . 'events' . DS . $id . '.jpg';
-		unlink($file);
+	
+	public function beforeDelete($cascade) {
+		if (!parent::beforeDelete($cascade)) {
+			return false;
+		}
+		$file = APP . 'webroot' . DS . 'img' . DS . 'events' . DS . $this->id . '.jpg';
+		if (file_exists($file)) {
+			unlink($file);
+		}
+		return true;
 	}
 		
 }
